+++ lib/cake_flutter.dart
@@ -7,6 +7,7 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_test/flutter_test.dart';
+
 export 'package:cake/cake.dart';
 
 part 'flutter_context.dart';
@@ -14,8 +15,10 @@ part 'flutter_expect.dart';
 part 'flutter_test_options.dart';
 part 'flutter_test_runner.dart';
 part 'index_options.dart';
+part 'root_widget.dart';
 part 'search.dart';
+part 'setup_settings.dart';
 part 'test_element_actions.dart';
-part 'test_element_wrapper_collection.dart';
 part 'test_element_wrapper.dart';
+part 'test_element_wrapper_collection.dart';
 part 'widget_tree.dart';